name reform
This was a doozy, with a big false start on an abstract LPContainer type. In the end we got a shaken-up reified with somewhat more coherence regarding variable ownership.

As a bonus, we also got a builder pattern for child naming, as well as progress toward a more coherent model of solver capabilities.