another big reform; hail permanent refactoring
Deep change: added a wrapper around the number type to LPAffExpr, so now we carry around the reified type with us. This lets us be way more generic in way more places, approaching total genericity

Also adds a builder-style pattern for LPNamed which makes writing constraints much, much shorter and prettier, in tandem with a plethora of helper methods on LPAffExpr

This genericity refactor actually caught a few bugs too, that's always a good sign.